requirements.txt
streamlit>=1.28.0
pandas>=1.3.0
python-dateutil>=2.8.0
pytest>=7.0.0
google-genai>=1.0.0
